+++ index.html
@@ -41,6 +41,15 @@
 
 
     </div>
-    <script src="./script.js"></script>
+    <script type="module" src="./script.js"></script>
+    <script type="module" src="./checkboxes.js"></script>
+    <script type="module" src="./combat.js"></script>
+    <script type="module" src="./interaction.js"></script>
+    <script type="module" src="./inventory.js"></script>
+    <script type="module" src="./location.js"></script>
+    <script type="module" src="./market.js"></script>
+    <script type="module" src="./player.js"></script>
+    <script type="module" src="./storage.js"></script>
+
 </body>
 </html>
