+++ interaction.js
@@ -1,0 +1,23 @@
+import './script.js';
+import './market.js';
+import './player.js';
+import './combat.js';
+import './storage.js';
+import './inventory.js';
+import './checkboxes.js';
+import './location.js';
+import './interaction.js';
+
+
+const button1 = document.querySelector('#button1');
+const button2 = document.querySelector('#button2');
+const button3 = document.querySelector('#button3');
+const button4 = document.querySelector('#button4');
+const buttonInv = document.querySelector('#buttonInv');
+const text    = document.querySelector('#text');
+const charismaText = document.querySelector('#charismaText');
+const healthText   = document.querySelector('#healthText');
+const goldText = document.querySelector('#goldText');
+const monsterStats = document.querySelector('#monsterStats');
+const monsterName = document.querySelector('#monsterName');
+const monsterHealthText = document.querySelector('#monsterHealth');
